pyproject.toml
[project]
name = "blog-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
    "requests>=2.32.4",
]

[tool.ruff]
extend-exclude = ['migrations', 'settings.py', 'manage.py']
line-length = 99

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR6301', 'PLC0415']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.7",
    "taskipy>=1.14.1",
]
